+++ docs/CHANGELOG.rst
@@ -31,6 +31,7 @@ Added
 - Add ``wgs-preprocess`` process
 - Add ``gatk-haplotypecaller-gvcf`` process
 - Add ``workflow-wgs-gvcf`` process
+- Add ``gatk-genotype-gvcfs`` process
 
 
 ===================